viewCreator更新作業
